requirements.txt
streamlit==1.37.0
ultralytics==8.3.34
opencv-python-headless==4.10.0.84
numpy==2.2.2
pillow==10.4.0
